+++ src/main/java/com/github/lmh01/mgt2mt/content/managed/ContentAdministrator.java
@@ -1,8 +1,10 @@
 package com.github.lmh01.mgt2mt.content.managed;
 
+import com.github.lmh01.mgt2mt.MadGamesTycoon2ModTool;
 import com.github.lmh01.mgt2mt.content.manager.*;
 import com.github.lmh01.mgt2mt.data_stream.UpdateChecker;
 import com.github.lmh01.mgt2mt.util.I18n;
+import com.github.lmh01.mgt2mt.util.OSType;
 import com.github.lmh01.mgt2mt.util.helper.DebugHelper;
 import com.github.lmh01.mgt2mt.util.helper.TextAreaHelper;
 import org.slf4j.Logger;
@@ -98,6 +100,10 @@ public class ContentAdministrator {
             } else {
                 TextAreaHelper.appendText(I18n.INSTANCE.get("warnMessage.integrityCheckFailed.textArea.mod_tool_not_compatible"));
                 TextAreaHelper.appendText(I18n.INSTANCE.get("warnMessage.integrityCheckFailed.textArea.checkGithub"));
+            }
+            if (MadGamesTycoon2ModTool.OS_TYPE.equals(OSType.LINUX)) {
+                TextAreaHelper.appendText(I18n.INSTANCE.get("warnMessage.integrityCheckFailed.textArea.linux"));
+                TextAreaHelper.appendText(I18n.INSTANCE.get("warnMessage.integrityCheckFailed.textArea.linux2"));
             }
         }
     }